[Working!!11! (confirned reel)) Removed Controller and now use DDPConnection as base class
