+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub mod controller;
+pub mod connection;
 pub mod error;
 pub mod packet;
 pub mod protocol;
